requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
Werkzeug==3.1.3
flask-sqlalchemy
flask-cors
flask-marshmallow
marshmallow-sqlalchemy
marshmallow
psycopg2-binary